requirements.txt
streamlit==1.49.1
pandas==2.3.1
pytest==8.4.1
uvicorn==0.35.0
mysql-connector-python==9.4.0
fastapi==0.116.1
requests==2.32.4
pydentic==0.0.1

